finalizado

+++ sitemaGerenciamentoTarefas/src/main/java/com/gustavo/sitemaGerenciamentoTarefas/service/TarefaService.java
@@ -1,8 +1,6 @@
 package com.gustavo.sitemaGerenciamentoTarefas.service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
@@ -15,7 +13,6 @@ import com.gustavo.sitemaGerenciamentoTarefas.repository.TarefaRepositoy;
 @Service
 public class TarefaService {
     
-    List<Tarefa> lista = new ArrayList<>();
     private final TarefaRepositoy repository;
 
     /* Construto */
@@ -53,7 +50,7 @@ public class TarefaService {
     //Salvar Tarefas
     public Tarefa salvar(TarefaRequest dto){
 
-        Tarefa p = new Tarefa();
+        Tarefa p = toEntity(dto);
         p.setNome(dto.getNome());
         p.setDescricao(dto.getDescricao());
         p.setConcluida(dto.getConcluida());
@@ -62,30 +59,27 @@ public class TarefaService {
 
     //Buscar por id
     public Tarefa buscarId(int id){
-        for(Tarefa f : lista){
-            if(f.getId() != id){
-                throw new TarefaNaoEncontradaException("Id não encontrado");
-            }
-        }
-        Optional<Tarefa> op = repository.findById(id);
-        return op.orElse(null); 
+        return repository.findById(id)
+        .orElseThrow(() -> new TarefaNaoEncontradaException("Id não encontrado"));
     }
 
     //Atualizar
     public Tarefa atualizar(int id, TarefaRequest dto){
-        for(Tarefa f : lista){
-            if(f.getId() == id){
-                f.setNome(dto.getNome());
-                f.setDescricao(dto.getDescricao());
-                f.setConcluida(dto.getConcluida());
-                return f;
-            }
-        }
-        throw new TarefaNaoEncontradaException("Id não encontrada");
+        Tarefa tarefa = repository.findById(id)
+        .orElseThrow(() -> new TarefaNaoEncontradaException("Id não encontrada"));
+
+        tarefa.setNome(dto.getNome());
+        tarefa.setDescricao(dto.getDescricao());
+        tarefa.setConcluida(dto.getConcluida());
+
+        return repository.save(tarefa);
     }
 
     //deletar
     public void deletar(int id){
+        if(!repository.existsById(id)){
+            throw new TarefaNaoEncontradaException("Id não encontrado");
+        }
         repository.deleteById(id);
     }
 
@@ -96,7 +90,7 @@ public class TarefaService {
 
     //Filtrar por titulo
     public List<Tarefa> filtrarTitulo(String text){
-        return repository.findByContainig(text);
+        return repository.findByContaining(text);
     }
 
 }
